workers/requirements.txt
boto3~=1.17
tweepy~=3.10
praw~=7.2
python-dotenv~=0.16
